+++ Pandora/Paradigm/Basis/Constant.hs
@@ -6,6 +6,14 @@ import Pandora.Pattern.Functor.Contravariant (Contravariant ((>$<)))
 import Pandora.Pattern.Functor.Invariant (Invariant (invmap))
 import Pandora.Pattern.Functor.Pointable (Pointable (point))
 import Pandora.Pattern.Functor.Traversable (Traversable ((->>)))
+import Pandora.Pattern.Object.Setoid (Setoid ((==)))
+import Pandora.Pattern.Object.Chain (Chain ((<=>)))
+import Pandora.Pattern.Object.Semigroup (Semigroup ((<>)))
+import Pandora.Pattern.Object.Monoid (Monoid (unit))
+import Pandora.Pattern.Object.Ringoid (Ringoid ((><)))
+import Pandora.Pattern.Object.Semilattice (Infimum ((/\)), Supremum ((\/)))
+import Pandora.Pattern.Object.Lattice (Lattice)
+import Pandora.Pattern.Object.Group (Group (inverse))
 
 newtype Constant a b = Constant a
 
@@ -20,3 +28,29 @@ instance Invariant (Constant a) where
 
 instance Traversable (Constant a) where
 	Constant x ->> _ = point (Constant x)
+
+instance Setoid a => Setoid (Constant a b) where
+	Constant x == Constant y = x == y
+
+instance Chain a => Chain (Constant a b) where
+	Constant x <=> Constant y = x <=> y
+
+instance Semigroup a => Semigroup (Constant a b) where
+	Constant x <> Constant y = Constant $ x <> y
+
+instance Monoid a => Monoid (Constant a b) where
+	 unit = Constant unit
+
+instance Ringoid a => Ringoid (Constant a b) where
+	Constant x >< Constant y = Constant $ x >< y
+
+instance Infimum a => Infimum (Constant a b) where
+	Constant x /\ Constant y = Constant $ x /\ y
+
+instance Supremum a => Supremum (Constant a b) where
+	Constant x \/ Constant y = Constant $ x \/ y
+
+instance Lattice a => Lattice (Constant a b) where
+
+instance Group a => Group (Constant a b) where
+	inverse (Constant x) = Constant $ inverse x